Menu page structure
Added the basic structure to the menu page including the checker-board layout.

Swapped the image and text in events and contacts sections on menu page to make it better suit the menu section.

Removed middle-footer and merged the content into lower-footer.

Adjusted the footer on all pages to make the elements sit in their intended places.