Mostly the index page.
Replaced some of the placeholder art.

Minor change to the wording of the info section.

Solved the mystery invisible bar on images: it was default padding in the img tag. Both margin and padding have been set to 0.

Added root text colours and a border standard.

Changed logo to be in em units.

Brain fried.