Half-way there?
Changed the hero image on each page.
Fixed the mess on index page that was created by css changes in an earlier branch.
Changes to the navbar and both logos (small and large) to use em and rem instead of vh.
Changed the height of the upper-footer.
Changed the methodology for sizing the lower footer, resulting in no visual change.
Changed most vh and vw units to %, and some (padding) to rem.