Last one for the day.
Resolved an issue where the menu would not fill the width of the screen at high resolutions.
Contact page renamed to contact us.
Resolved an issue causing the map section of the contact us page to display incorrectly.
Many changes to CSS which will have surely broken something that will need to be fixed.
Media queries added for link to css.
style.css renamed to desktop.css
Link added to the header, and commented out, for the mobile.css that will come later.
Navbar was rebuilt.
The day is done and so am I.