The gnomes are stealing my sanity.
Links added to all the images on the credits page.
Removed the placeholder fancy pants logo from the project files.
Fixed a cropping issue with the showcase image on the credits page.
Updated menu and find us links to take the user to the applicable section of the linked page.
Added mobile.css and made first attempts at alterations.
So many alterations to both CSS and HTML to make this work.